Updated due to discovered bug
There was a reported instance of the item code "uar" matching Russet Armor. Russet Armor appears to match correctly with its own item code of "xpl". It is currently unknown if Sacred Armor correctly matches with "uar", but no evidence has been provided with it not working with it yet. So this update is solely to include a workaround to exclude Russet Armor from matching in the two lines where "uar" is used.