New Main Filter
Due to all the changes that occurred with patch #20 - Perlite, the two previous versions will no longer be maintained. The item level is now shown by default through the item descriptions. In addition the file no longer has to be named "item.filter", so the filter is now just named "Spam's Filter.filter" so it can be easily distinguished from other filters in the drop down list available.